project/supervision_handler/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
eventlet==0.36.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.9.0
